refactor(tests): Add UT for livestream
refactor(gh-action): Add --update flag

refactor(aws-cli): Remove AWS cli installation
